+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 - Tags: new auto-tags named 'bPinnedFirst' and 'bPinnedLast' which force showing the tagged playlists at top/bottom of the list, for any sorting/filtering combination.
 - UI: script will be completely disabled after installing for the first time until user clicks on the setup button (which will initiate a serie of popups). This is a workaround for a [SMP bug](https://github.com/TheQwertiest/foo_spider_monkey_panel/issues/210) during the installation process.
 - UI: added manual sorting. It must be enabled on the sorting button and works independently to the automatic sorting methods. Playlist can be reordered by drag n' dropping within the panel or using the sorting submenu. It also works with multiple selection. Pinned playlists are ignored in this mode. Sorting indexes are stored at '.\js_data\'.
+- UI: configurable columns to display playlist metadata. Width, metadata, font size, order, num of columns, borders and colors can be set. In the case of playlist's size, a configurable unit may be set. The entire feature may be disabled using the configurable UI elements submenu.
 - UI: added sorting by creation date.
 - UI: added configurable R. click mouse actions (like the previous L. click and Middle click ones).
 - UI: added configurable timer for [double clicking](https://en.wikipedia.org/wiki/Double-click).
@@ -57,6 +58,7 @@
 - Keyboard shortcuts: pressing F10 will open the settings menu or the list menu (when also pressing shift).
 - Keyboard shortcuts: pressing F11 will open the documentation (pdf).
 - Keyboard shortcuts: pressing F11 will open playlists tracked folder.
+- Keyboard shortcuts: pressing 'º', '\' or 'Numpad /' will globally hide/show the playlist's metadata columns.
 - Playlist formats: creation and last modified date values are now cached for UI-only playlists and Auto-Playlists. To be used along the new date sorting options.
 - Playlist formats: UI-only playlists metadata (tag, category and tracks' tags) is now editable (and cached between sessions). This allows for sorting and categorization of UI-only playlists.
 - ListenBrainz: playlists can now be exported to Spotify (when exporting to ListenBrainz). Requires Spotify's service to be connected to the user profile, and \'Play music on ListenBrainz\' [enabled](https://listenbrainz.org/profile/music-services/details/).
@@ -74,6 +76,7 @@
 - XSPF: optimizations finding tracks on library.
 - Playlist formats: creation and last modified date values are calculated when loading the playlist files, and cached during the entire session. Sorting by date is now much faster.
 - Drag and Drop: sending tracks to blank space or list menu button (plus shape) will create a new playlist.
+- UI: 'show size' option is now disabled by default. i.e. columns offer the same functionality now.
 - UI: renamed some menu entries for playlist creation.
 - UI: R. clicking now opens the contextual menu for the selected playlists (previously on L. Click). Shift + R. Clicking opens the playlist's items native contextual menu.
 - UI: L. single click action is now configurable. 
@@ -95,6 +98,7 @@
 - UI: list view not maintaining current position at some instances. 'Online sync...'.
 - UI: focus was lost in some cases when editing an UI-only playlist.
 - UI: scrollbar not working properly when list exceeded current window size by one row.
+- UI: multiple minor fixes to menu entry checks (the tick shown if setting is enabled).
 - Auto-Functions: fix 'bAutoLoad' improper loading in some cases.
 - Error while saving a playlist which was the first one on UI.